Convert to double just in case in update

+++ lib/src/application/update_loop.dart
@@ -105,7 +105,7 @@ class UpdateLoop {
     // now() within the application.
     time = html.window.performance.now();
 
-    Time.update(time);
+    Time.update(time.toDouble());
 
     if (onFrame != null) {
       onFrame(new Time());
